lambda/functions/notesOcrJob/requirements.txt
google-cloud-documentai==2.27.0
boto3==1.34.99